requirements.txt
streamlit==1.32.0
requests==2.31.0
python-dotenv==1.0.0
transformers==4.45.0
torch==2.8.0
